Moved file positions
